Introduced httpRedirectHtml()
While doing so fixed HTML bugs predating the commit "Extracted
ZimDumper::writeHttpRedirect()":
1. No opening `<html>` tag
2. A second `<head>` tag was used instead of the closing `</head>` tag
