+++ src/tools.cpp
@@ -25,6 +25,7 @@
 #include <sys/stat.h>
 #include <cerrno>
 #include <fstream>
+#include <iomanip>
 #include <iostream>
 #include <sstream>
 #include <stdexcept>
@@ -33,6 +34,7 @@
 #include <unistd.h>
 #include <algorithm>
 #include <regex>
+#include <array>
 
 #ifdef _WIN32
 #define SEPARATOR "\\"
@@ -450,3 +452,63 @@ UriKind html_link::detectUriKind(const std::string& input_string)
     return specialUriSchemeKind(scheme);
 }
 
+namespace
+{
+
+static bool isReservedUrlChar(const char c)
+{
+    constexpr std::array<char, 10> reserved = {{';', ',', '?', ':',
+                                               '@', '&', '=', '+', '$' }};
+
+    return std::any_of(reserved.begin(), reserved.end(),
+                       [&c] (const char &elem) { return elem == c; } );
+}
+
+bool needsEscape(const char c, const bool encodeReserved)
+{
+  if ((c >= 'a' && c <= 'z') || (c >= 'A' && c <= 'Z') || (c >= '0' && c <= '9'))
+    return false;
+
+  if (isReservedUrlChar(c))
+    return encodeReserved;
+
+  constexpr std::array<char, 10> noNeedEscape = {{'-', '_', '.', '!', '~',
+                                                '*', '\'', '(', ')', '/' }};
+
+  return not std::any_of(noNeedEscape.begin(), noNeedEscape.end(),
+                         [&c] (const char &elem) { return elem == c; } );
+}
+
+std::string urlEncode(const std::string& value, bool encodeReserved)
+{
+  std::ostringstream os;
+  os << std::hex << std::uppercase;
+  for (std::string::const_iterator it = value.begin();
+       it != value.end();
+       ++it) {
+    if (!needsEscape(*it, encodeReserved)) {
+      os << *it;
+    } else {
+      os << '%' << std::setw(2) << static_cast<unsigned int>(static_cast<unsigned char>(*it));
+    }
+  }
+  return os.str();
+}
+
+} // unnamed namespace
+
+std::string httpRedirectHtml(const std::string& redirectUrl)
+{
+    const auto encodedurl = urlEncode(redirectUrl, true);
+    std::ostringstream ss;
+
+    ss << "<!DOCTYPE html>"
+          "<html>"
+          "<head>"
+          "<meta http-equiv=\"Content-Type\" content=\"text/html; charset=utf-8\" />"
+          "<meta http-equiv=\"refresh\" content=\"0;url=" + encodedurl + "\" />"
+          "</head>"
+          "<body></body>"
+          "</html>";
+    return ss.str();
+}
